C8a: decompose wasm.py into wasm/ package (#100)
Split the 2,344-line wasm.py monolith into a vera/wasm/ package with
7 submodules using the same mixin-based pattern as the checker/ package.

Package layout:
  __init__.py    (6 lines)   — re-exports 4 public API symbols
  helpers.py   (211 lines)   — WasmSlotEnv, StringPool, module-level helpers
  context.py   (369 lines)   — composed WasmContext class + dispatcher
  inference.py (527 lines)   — InferenceMixin: type inference + utilities
  operators.py (430 lines)   — OperatorsMixin: binary/unary ops, control flow
  calls.py     (223 lines)   — CallsMixin: function calls, effect handlers
  closures.py  (248 lines)   — ClosuresMixin: closures, free-var analysis
  data.py      (460 lines)   — DataMixin: constructors, match, arrays

No test changes. All 951 tests pass, mypy clean, all 14 examples pass.
