+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "vera"
-version = "0.0.40"
+version = "0.0.41"
 description = "Vera: a programming language designed for LLMs, with full contracts, algebraic effects, and typed slot references"
 readme = "README.md"
 license = "MIT"
@@ -86,5 +86,19 @@ module = [
     "vera.checker.expressions",
     "vera.checker.calls",
     "vera.checker.control",
+]
+disable_error_code = ["attr-defined", "no-any-return", "has-type"]
+
+[[tool.mypy.overrides]]
+# WASM mixin classes reference attributes (self.string_pool, self.alloc_local,
+# etc.) that live on the composed WasmContext class in context.py.  Same
+# mixin pattern as the checker modules above.
+module = [
+    "vera.wasm.context",
+    "vera.wasm.inference",
+    "vera.wasm.operators",
+    "vera.wasm.calls",
+    "vera.wasm.closures",
+    "vera.wasm.data",
 ]
 disable_error_code = ["attr-defined", "no-any-return", "has-type"]
