+++ vera/wasm/__init__.py
@@ -1,0 +1,6 @@
+"""Vera WASM translation layer — AST to WAT bridge."""
+
+from vera.wasm.context import WasmContext
+from vera.wasm.helpers import StringPool, WasmSlotEnv, wasm_type
+
+__all__ = ["StringPool", "WasmContext", "WasmSlotEnv", "wasm_type"]
