+++ vera/wasm/calls.py
@@ -1,0 +1,223 @@
+"""Function call and effect handler translation mixin for WasmContext."""
+
+from __future__ import annotations
+
+from vera import ast
+from vera.wasm.helpers import WasmSlotEnv
+
+
+class CallsMixin:
+    """Methods for translating function calls and effect handlers to WASM."""
+
+    def _translate_call(
+        self, call: ast.FnCall, env: WasmSlotEnv
+    ) -> list[str] | None:
+        """Translate a function call to WASM call instruction.
+
+        If the call name matches an effect operation (e.g. get/put for
+        State<T>), redirects to the corresponding host import.
+        """
+        # Built-in: length(array) → Int
+        if call.name == "length" and len(call.args) == 1:
+            return self._translate_length(call.args[0], env)
+
+        # Check if this is a closure application: apply_fn(closure, args...)
+        if call.name == "apply_fn" and len(call.args) >= 2:
+            return self._translate_apply_fn(call, env)
+
+        # Check if this is an effect operation (e.g. get/put)
+        if call.name in self._effect_ops:
+            import_name, _is_void = self._effect_ops[call.name]
+            instructions: list[str] = []
+            for arg in call.args:
+                arg_instrs = self.translate_expr(arg, env)
+                if arg_instrs is None:
+                    return None
+                instructions.extend(arg_instrs)
+            instructions.append(f"call {import_name}")
+            return instructions
+
+        # Resolve call target — rewrite generic calls to mangled names
+        call_target = call.name
+        if call.name in self._generic_fn_info:
+            resolved = self._resolve_generic_call(call)
+            if resolved is not None:
+                call_target = resolved
+
+        # Guard rail: reject calls to functions not defined in this module
+        if (self._known_fns
+                and call_target not in self._known_fns
+                and call_target not in self._ctor_layouts):
+            return None
+
+        # Regular function call
+        instructions = []
+        for arg in call.args:
+            arg_instrs = self.translate_expr(arg, env)
+            if arg_instrs is None:
+                return None
+            instructions.extend(arg_instrs)
+        instructions.append(f"call ${call_target}")
+        return instructions
+
+    def _translate_qualified_call(
+        self, call: ast.QualifiedCall, env: WasmSlotEnv
+    ) -> list[str] | None:
+        """Translate a qualified call (e.g. IO.print) to host import call."""
+        # Only IO effect operations are supported in C5
+        instructions: list[str] = []
+        for arg in call.args:
+            arg_instrs = self.translate_expr(arg, env)
+            if arg_instrs is None:
+                return None
+            instructions.extend(arg_instrs)
+        instructions.append(f"call $vera.{call.name}")
+        return instructions
+
+    def _resolve_generic_call(self, call: ast.FnCall) -> str | None:
+        """Resolve a call to a generic function to its mangled name.
+
+        Infers concrete type variable bindings from the call's argument
+        expressions, then produces the mangled name like 'identity$Int'.
+        Returns None if type inference fails.
+        """
+        forall_vars, param_types = self._generic_fn_info[call.name]
+        mapping: dict[str, str] = {}
+
+        for param_te, arg in zip(param_types, call.args):
+            self._unify_param_arg_wasm(param_te, arg, forall_vars, mapping)
+
+        # Build mangled name
+        parts = []
+        for tv in forall_vars:
+            if tv not in mapping:
+                return None
+            parts.append(mapping[tv])
+        return f"{call.name}${'_'.join(parts)}"
+
+    def _unify_param_arg_wasm(
+        self,
+        param_te: ast.TypeExpr,
+        arg: ast.Expr,
+        forall_vars: tuple[str, ...],
+        mapping: dict[str, str],
+    ) -> None:
+        """Unify a parameter TypeExpr against an argument to bind type vars.
+
+        Mirrors CodeGenerator._unify_param_arg for use during WASM
+        translation.
+        """
+        if isinstance(param_te, ast.RefinementType):
+            self._unify_param_arg_wasm(
+                param_te.base_type, arg, forall_vars, mapping,
+            )
+            return
+
+        if not isinstance(param_te, ast.NamedType):
+            return
+
+        if param_te.name in forall_vars:
+            vera_type = self._infer_vera_type(arg)
+            if vera_type and param_te.name not in mapping:
+                mapping[param_te.name] = vera_type
+            return
+
+        # Parameterized type like Option<T>
+        if param_te.type_args:
+            arg_info = self._get_arg_type_info_wasm(arg)
+            if arg_info and arg_info[0] == param_te.name:
+                for param_ta, arg_ta_name in zip(
+                    param_te.type_args, arg_info[1]
+                ):
+                    if (isinstance(param_ta, ast.NamedType)
+                            and param_ta.name in forall_vars
+                            and param_ta.name not in mapping):
+                        mapping[param_ta.name] = arg_ta_name
+
+    def _translate_length(
+        self, arg: ast.Expr, env: WasmSlotEnv,
+    ) -> list[str] | None:
+        """Translate length(array) → Int (i64).
+
+        Evaluates the array → (ptr, len), drops ptr, extends len to i64.
+        """
+        arg_instrs = self.translate_expr(arg, env)
+        if arg_instrs is None:
+            return None
+        tmp_len = self.alloc_local("i32")
+        instructions: list[str] = []
+        instructions.extend(arg_instrs)
+        # Stack has (ptr, len); save len, drop ptr
+        instructions.append(f"local.set {tmp_len}")
+        instructions.append("drop")
+        instructions.append(f"local.get {tmp_len}")
+        instructions.append("i64.extend_i32_u")
+        return instructions
+
+    def _translate_handle_expr(
+        self, expr: ast.HandleExpr, env: WasmSlotEnv,
+    ) -> list[str] | None:
+        """Translate a handle expression to WASM.
+
+        Currently supports State<T> handlers via host imports.
+        Other handler types cause the function to be skipped.
+        """
+        effect = expr.effect
+        if not isinstance(effect, ast.EffectRef):
+            return None
+
+        if effect.name == "State" and effect.type_args and len(effect.type_args) == 1:
+            return self._translate_handle_state(expr, env)
+
+        # Unsupported handler type
+        return None
+
+    def _translate_handle_state(
+        self, expr: ast.HandleExpr, env: WasmSlotEnv,
+    ) -> list[str] | None:
+        """Translate handle[State<T>](@T = init) { ... } in { body }.
+
+        Compiles by:
+        1. Evaluating init_expr and calling state_put_T to set initial state
+        2. Temporarily injecting get/put effect ops for the body
+        3. Compiling the body with these ops active
+        4. Restoring the previous effect ops
+        """
+        assert isinstance(expr.effect, ast.EffectRef)
+        type_arg = expr.effect.type_args[0]  # type: ignore[index]
+        if isinstance(type_arg, ast.NamedType):
+            type_name = type_arg.name
+        else:
+            return None
+
+        wasm_type = self._type_name_to_wasm(type_name)
+        put_import = f"$vera.state_put_{type_name}"
+        get_import = f"$vera.state_get_{type_name}"
+
+        instructions: list[str] = []
+
+        # 1. Initialize state: compile init_expr, call state_put
+        if expr.state is not None:
+            init_instrs = self.translate_expr(expr.state.init_expr, env)
+            if init_instrs is None:
+                return None
+            instructions.extend(init_instrs)
+            instructions.append(f"call {put_import}")
+        # If no state clause, state starts at default (0)
+
+        # 2. Save current effect_ops and inject handler ops
+        saved_ops = dict(self._effect_ops)
+        self._effect_ops["get"] = (get_import, False)
+        self._effect_ops["put"] = (put_import, True)
+
+        # 3. Compile handler body
+        body_instrs = self.translate_block(expr.body, env)
+
+        # 4. Restore effect_ops
+        self._effect_ops = saved_ops
+
+        if body_instrs is None:
+            return None
+
+        instructions.extend(body_instrs)
+        return instructions
